backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests>=2.32.4",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "python-dotenv>=0.21.0",
]

[tool.pytest.ini_options]
pythonpath = [
  "."
]
